# src/Nox.ClientApi.Tests/ServiceMetadata/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nox.ClientApi.Tests",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Stores": {
      "get": {
        "tags": [
          "Stores"
        ],
        "operationId": "api/Stores",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json;odata.metadata=minimal;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=minimal": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=full": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.metadata=none": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "application/octet-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stores"
        ],
        "operationId": "api/Stores",
        "requestBody": {
          "content": {
            "application/json;odata.metadata=minimal;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=minimal": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=full": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.metadata=none": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StoreCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json;odata.metadata=minimal;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=minimal": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=full": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.metadata=none": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Stores/{key}/EmailAddress": {
      "delete": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountryCode": {
        "enum": [
          4,
          8,
          10,
          12,
          16,
          20,
          24,
          28,
          31,
          32,
          36,
          40,
          44,
          48,
          50,
          51,
          52,
          56,
          60,
          64,
          68,
          70,
          72,
          74,
          76,
          84,
          86,
          90,
          92,
          96,
          100,
          104,
          108,
          112,
          116,
          120,
          124,
          132,
          136,
          140,
          144,
          148,
          152,
          156,
          158,
          162,
          166,
          170,
          174,
          175,
          178,
          180,
          184,
          188,
          191,
          192,
          196,
          203,
          204,
          208,
          212,
          214,
          218,
          222,
          226,
          231,
          232,
          233,
          234,
          238,
          239,
          242,
          246,
          248,
          250,
          254,
          258,
          260,
          262,
          266,
          268,
          270,
          275,
          276,
          288,
          292,
          296,
          300,
          304,
          308,
          312,
          316,
          320,
          324,
          328,
          332,
          334,
          336,
          340,
          344,
          348,
          352,
          356,
          360,
          364,
          368,
          372,
          376,
          380,
          384,
          388,
          392,
          398,
          400,
          404,
          408,
          410,
          414,
          417,
          418,
          422,
          426,
          428,
          430,
          434,
          438,
          440,
          442,
          446,
          450,
          454,
          458,
          462,
          466,
          470,
          474,
          478,
          480,
          484,
          492,
          496,
          498,
          499,
          500,
          504,
          508,
          512,
          516,
          520,
          524,
          528,
          531,
          533,
          534,
          535,
          540,
          548,
          554,
          558,
          562,
          566,
          570,
          574,
          578,
          580,
          581,
          583,
          584,
          585,
          586,
          591,
          598,
          600,
          604,
          608,
          612,
          616,
          620,
          624,
          626,
          630,
          634,
          638,
          642,
          643,
          646,
          652,
          654,
          659,
          660,
          662,
          663,
          666,
          670,
          674,
          678,
          682,
          686,
          688,
          690,
          694,
          702,
          703,
          704,
          705,
          706,
          710,
          716,
          724,
          728,
          729,
          732,
          740,
          744,
          748,
          752,
          756,
          760,
          762,
          764,
          768,
          772,
          776,
          780,
          784,
          788,
          792,
          795,
          796,
          798,
          800,
          804,
          807,
          818,
          826,
          831,
          832,
          833,
          834,
          840,
          850,
          854,
          858,
          860,
          862,
          876,
          882,
          887,
          894
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "EmailAddressCreateDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EmailAddressDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LatLongDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "StoreCreateDto": {
        "required": [
          "address",
          "isTemporary",
          "location",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/StreetAddressDto"
          },
          "location": {
            "$ref": "#/components/schemas/LatLongDto"
          },
          "ownershipId": {
            "type": "string",
            "nullable": true
          },
          "verifiedEmails": {
            "$ref": "#/components/schemas/EmailAddressCreateDto"
          },
          "isTemporary": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "StoreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/StreetAddressDto"
          },
          "location": {
            "$ref": "#/components/schemas/LatLongDto"
          },
          "ownershipId": {
            "type": "string",
            "nullable": true
          },
          "ownership": {
            "$ref": "#/components/schemas/StoreOwnerDto"
          },
          "verifiedEmails": {
            "$ref": "#/components/schemas/EmailAddressDto"
          },
          "deletedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "@odata.etag": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "StoreOwnerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "temporaryOwnerName": {
            "type": "string",
            "nullable": true
          },
          "vatNumber": {
            "$ref": "#/components/schemas/VatNumberDto"
          },
          "streetAddress": {
            "$ref": "#/components/schemas/StreetAddressDto"
          },
          "localGreeting": {
            "$ref": "#/components/schemas/TranslatedTextDto"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "stores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoreDto"
            },
            "nullable": true
          },
          "deletedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "@odata.etag": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "StreetAddressDto": {
        "type": "object",
        "properties": {
          "streetNumber": {
            "type": "string",
            "nullable": true
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "route": {
            "type": "string",
            "nullable": true
          },
          "locality": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "administrativeArea1": {
            "type": "string",
            "nullable": true
          },
          "administrativeArea2": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "countryId": {
            "$ref": "#/components/schemas/CountryCode"
          }
        },
        "additionalProperties": false
      },
...
      "TranslatedTextDto": {
        "type": "object",
        "properties": {
          "cultureCode": {
            "type": "string",
            "nullable": true
          },
          "phrase": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VatNumberDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "$ref": "#/components/schemas/CountryCode"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}